Flytter kode fra forskyvVilkårResultater inn til ForskyvVilkårFørFebruar2025
